# Undo changes made in PR #2218 and optimize copying query results