--- a/src/Microsoft.SqlTools.ServiceLayer/QueryExecution/QueryExecutionService.cs
+++ b/src/Microsoft.SqlTools.ServiceLayer/QueryExecution/QueryExecutionService.cs
@@ -29,6 +29,7 @@ using Microsoft.SqlTools.ServiceLayer.SqlContext;
 using Microsoft.SqlTools.ServiceLayer.Workspace.Contracts;
 using Microsoft.SqlTools.ServiceLayer.Workspace;
 using Microsoft.SqlTools.Utility;
+using TextCopy;
 
 namespace Microsoft.SqlTools.ServiceLayer.QueryExecution
 {
@@ -206,6 +207,7 @@ namespace Microsoft.SqlTools.ServiceLayer.QueryExecution
             serviceHost.SetRequestHandler(SimpleExecuteRequest.Type, HandleSimpleExecuteRequest, true);
             serviceHost.SetRequestHandler(QueryExecutionOptionsRequest.Type, HandleQueryExecutionOptionsRequest, true);
             serviceHost.SetRequestHandler(CopyResultsRequest.Type, HandleCopyResultsRequest, true);
+            serviceHost.SetEventHandler(CopyResultsToClipboardEvent.Type, HandleCopyResultsToClipboardNotification, true);
 
             // Register the file open update handler
             WorkspaceService<SqlToolsSettings>.Instance.RegisterTextDocCloseCallback(HandleDidCloseTextDocumentNotification);
@@ -778,6 +780,27 @@ namespace Microsoft.SqlTools.ServiceLayer.QueryExecution
         /// </summary>
         internal async Task HandleCopyResultsRequest(CopyResultsRequestParams requestParams, RequestContext<CopyResultsRequestResult> requestContext)
         {
+            StringBuilder builder = await GetStringBuilderToCopyQueryResults(requestParams);
+
+            CopyResultsRequestResult result = new CopyResultsRequestResult
+            {
+                Results = builder.ToString()
+            };
+            await requestContext.SendResult(result);
+        }
+
+        /// <summary>
+        /// Handles copying results to the clipboard.
+        /// </summary>
+        internal async Task HandleCopyResultsToClipboardNotification(CopyResultsRequestParams requestParams, EventContext eventContext)
+        {
+            StringBuilder builder = await GetStringBuilderToCopyQueryResults(requestParams);
+
+            await ClipboardService.SetTextAsync(builder.ToString());
+        }
+
+        private async Task<StringBuilder> GetStringBuilderToCopyQueryResults(CopyResultsRequestParams requestParams)
+        {
             var valueSeparator = "\t";
             var columnRanges = this.MergeRanges(requestParams.Selections.Select(selection => new Range() { Start = selection.FromColumn, End = selection.ToColumn }).ToList());
             var rowRanges = this.MergeRanges(requestParams.Selections.Select(selection => new Range() { Start = selection.FromRow, End = selection.ToRow }).ToList());
@@ -871,11 +894,8 @@ namespace Microsoft.SqlTools.ServiceLayer.QueryExecution
                     pageStartRowIndex += rowsToFetch;
                 } while (pageStartRowIndex < rowRange.End);
             }
-            CopyResultsRequestResult result = new CopyResultsRequestResult
-            {
-                Results = builder.ToString()
-            };
-            await requestContext.SendResult(result);
+
+            return builder;
         }
 
         #endregion