This PR is a part of https://github.com/microsoft/azuredatastudio/issues/25632

This PR undoes #2218, which removed functionality to copy query results from SQL Tools Service (STS). This PR brings back functionality to copy results into the clipboard from STS. This PR also removes a bottleneck that was making it take a long time to copy query results because the `ToString`  method was being repeatedly invoked on the string builder that is assembling the query results string that is added to the clipboard.


After optimizing the copy endpoint to stop repeatedly invoking `ToString()` on the `StringBuilder` the time needed to copy query results became significantly faster:

Before removing the repeated `ToString()` calls:
Copying 25,000 rows took 7,982 ms
Copying 50,000 rows took 34,863 ms
Copying 75,000 rows took 76,430 ms
Copying 100,000 rows took 163,462 ms

After removing repeated `ToString()` calls:
Copying 25,000 rows took 1.0278617 ms
Copying 50,000 rows took 1.9388735 ms
Copying 75,000 rows took 2.8113253 ms
Copying 100,000 rows took 3.8120204 ms