@lewis-sanchez 2024-08-26
> @lewis-sanchez is it possible to do some basic benchmarking for this change? For example, with some very large resultssets how long does copy take with & without this change on various environments (specifically Linux and Windows). With all the previous PRs in this area linked above, I want to make sure we're address the issue reported in the most recent bug.

@kburtram, yes, it should be possible to do some benchmarking for this change.